adds function and removes class from top menu
