+++ ui/src/components/BottomMenu.js
@@ -13,10 +13,6 @@ import Drawer from '@material-ui/core/Drawer';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import ExpandLessIcon from '@material-ui/icons/ExpandLess';
 import Divider from '@material-ui/core/Divider';
-// import List from '@material-ui/core/List';
-// import ListItem from '@material-ui/core/ListItem';
-// import ListItemText from '@material-ui/core/ListItemText';
-// import ListSubheader from '@material-ui/core/ListSubheader';
 
 import {drawerToggle} from '../actions/NavigationActions'
 
@@ -53,13 +49,9 @@ const styles = theme => ({
         marginLeft: 0,
         marginRight: 28,
     },
-    // toolbar: {
-    //     alignItems: 'center',
-    //     justifyContent: 'space-between',
+    // hide: {
+    //     display: 'none',
     // },
-    hide: {
-        display: 'none',
-    },
     drawer: {
         width: drawerWidth,
         flexShrink: 0,
@@ -108,25 +100,10 @@ class BottomMenu extends React.Component{
 
         return (
             <div className={classes.root}>
-                {/* <Paper square className={classes.paper}>
-                    <Typography className={classes.text} variant="h5" gutterBottom>
-                    Controls
-                    </Typography>
-                    <Button variant="contained" className={classes.button}>
-                    START
-                    </Button>
-                    <Button variant="contained" color="primary" className={classes.button}>
-                    STOP
-                    </Button>
-                </Paper> */}
-
-
                 <AppBar position="fixed" color="primary" className={classes.appBar} style={{flexDirection: 'row'}}>
-                    {/* <Toolbar className={classes.toolbar}> */}
                     <IconButton onClick={() => { this.props.drawerToggle(!this.props.drawerOpen)}} className={classes.menuButton} color="inherit" aria-label="Open drawer">
                         <MenuIcon />
                     </IconButton>
-                    {/* </Toolbar> */}
                 </AppBar>
 
                 <Drawer
@@ -152,17 +129,6 @@ class BottomMenu extends React.Component{
                         </IconButton>
                     </div>
                     {/* <Divider /> */}
-                    {/* <List>
-                        {['Local Node Simulation', 'Remote Node Simulation', 'Local Data Center Simulation', 'Custom Event Simulation'].map((text, index) => (
-                            <ListItem button key={text} onClick={(e) => { this.props.changeScreen(text)}}>
-                            <ListItemIcon>
-                                {index === 0 ? <Filter1 /> : index === 1 ? <Filter2 /> : index === 2 ? <Filter3 /> : index === 3 ? <Filter4/> : <Filter5/>}
-                            </ListItemIcon>
-                            <ListItemText primary={text} />
-                            </ListItem>
-                        ))}
-                    </List> */}
-
                     <Paper square className={classes.paper}>
                         <Typography className={classes.text} variant="h5" gutterBottom>
                         Controls
