--- ui/src/components/Menu.js
@@ -1,241 +1,0 @@
-import React from 'react';
-import { connect } from 'react-redux';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import IconButton from '@material-ui/core/IconButton';
-import Typography from '@material-ui/core/Typography';
-import { fade } from '@material-ui/core/styles/colorManipulator';
-import { withStyles } from '@material-ui/core/styles';
-import MenuIcon from '@material-ui/icons/Menu';
-import SearchIcon from '@material-ui/icons/Search';
-import Drawer from '@material-ui/core/Drawer';
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import ChevronRightIcon from '@material-ui/icons/ChevronRight';
-import Divider from '@material-ui/core/Divider';
-import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
-import ListItemText from '@material-ui/core/ListItemText';
-import Filter1 from '@material-ui/icons/Filter1';
-import Filter2 from '@material-ui/icons/Filter2';
-import Filter3 from '@material-ui/icons/Filter3';
-import Filter4 from '@material-ui/icons/Filter4';
-import Filter5 from '@material-ui/icons/Filter5';
-
-import {drawerToggle, changeScreen} from '../actions/NavigationActions'
-
-import classNames from 'classnames';
-
-const drawerWidth = 300;
-
-const styles = theme => ({
-  root: {
-    width: '100%',
-  },
-  grow: {
-    flexGrow: 1,
-  },
-  appBar: {
-    zIndex: theme.zIndex.drawer + 1,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-  },
-  appBarShift: {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  },
-  menuButton: {
-    marginLeft: -12,
-    marginRight: 20,
-  },
-  title: {
-    display: 'none',
-    [theme.breakpoints.up('sm')]: {
-      display: 'block',
-    },
-  },
-  hide: {
-    display: 'none',
-  },
-  drawer: {
-    width: drawerWidth,
-    flexShrink: 0,
-    whiteSpace: 'nowrap',
-  },
-  drawerOpen: {
-    width: drawerWidth,
-    transition: theme.transitions.create('width', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  },
-  drawerClose: {
-    transition: theme.transitions.create('width', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-    overflowX: 'hidden',
-    width: theme.spacing.unit * 7 + 1,
-    [theme.breakpoints.up('sm')]: {
-      width: theme.spacing.unit * 9 + 1,
-    },
-  },
-  toolbar: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'flex-end',
-    padding: '0 8px',
-    ...theme.mixins.toolbar,
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing.unit * 3,
-  },
-  search: {
-    position: 'relative',
-    borderRadius: theme.shape.borderRadius,
-    backgroundColor: fade(theme.palette.common.white, 0.15),
-    '&:hover': {
-      backgroundColor: fade(theme.palette.common.white, 0.25),
-    },
-    marginLeft: 0,
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      marginLeft: theme.spacing.unit,
-      width: 'auto',
-    },
-  },
-  searchIcon: {
-    width: theme.spacing.unit * 9,
-    height: '100%',
-    position: 'absolute',
-    pointerEvents: 'none',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  inputRoot: {
-    color: 'inherit',
-    width: '100%',
-  },
-  inputInput: {
-    paddingTop: theme.spacing.unit,
-    paddingRight: theme.spacing.unit,
-    paddingBottom: theme.spacing.unit,
-    paddingLeft: theme.spacing.unit * 10,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      width: 120,
-      '&:focus': {
-        width: 200,
-      },
-    },
-  },
-});
-
-
-class Menu extends React.Component{
-  render({classes} = this.props){
-
-    return (
-      <div className={classes.root}>
-        <AppBar position="static">
-          <Toolbar>
-            {/* <IconButton onClick={() => { this.props.drawerToggle(!this.props.drawerOpen)}} className={classes.menuButton} color="inherit" aria-label="Open drawer">
-              <MenuIcon />
-            </IconButton> */}
-            <Typography className={classes.title} variant="h6" color="inherit" noWrap>
-              CHAOS IN THE CLOUDS
-              {/* CHAOS IN THE CLOUDS:  {this.props.page} */}
-            </Typography>
-            <div className={classes.grow} />
-            <div>
-                  <img style={{height: '80px'}}src={require('../images/logo.png')} />
-            </div>
-            {/* <div className={classes.search}>
-              <div className={classes.searchIcon}>
-              <SearchIcon />
-              </div>
-              <InputBase
-              placeholder="Search…"
-              classes={{
-                root: classes.inputRoot,
-                input: classes.inputInput,
-              }}
-              />
-            </div> */}
-          </Toolbar>
-        </AppBar>
-          {/* <Drawer
-          variant="permanent"
-          className={classNames(classes.drawer, {
-            [classes.drawerOpen]: this.props.drawerOpen,
-            [classes.drawerClose]: !this.props.drawerOpen,
-          })}
-          classes={{
-            paper: classNames({
-              [classes.drawerOpen]: this.props.drawerOpen,
-              [classes.drawerClose]: !this.props.drawerOpen,
-            }),
-          }}
-          open={this.props.drawerOpen}
-          >
-          <div className={classes.toolbar}>
-            <IconButton onClick={() => { this.props.drawerToggle(!this.props.drawerOpen)}}>
-              { !this.props.drawerOpen ? <ChevronRightIcon /> : <ChevronLeftIcon />}
-            </IconButton>
-          </div>
-          <Divider />
-            <List>
-              {['Local Node Simulation', 'Remote Node Simulation', 'Local Data Center Simulation', 'Custom Event Simulation'].map((text, index) => (
-                <ListItem button key={text} onClick={(e) => { this.props.changeScreen(text)}}>
-                  <ListItemIcon>
-                    {index === 0 ? <Filter1 /> : index === 1 ? <Filter2 /> : index === 2 ? <Filter3 /> : index === 3 ? <Filter4/> : <Filter5/>}
-                  </ListItemIcon>
-                  <ListItemText primary={text} />
-                </ListItem>
-              ))}
-            </List>
-        </Drawer> */}
-      </div>
-    );
-  }
-
-  componentDidMount(){
-    this.props.init()
-  }
-}
-
-const mapStateToProps = (state, ownProps) => {
-  return {
-    drawerOpen: state.NavigationReducer.drawerOpen,
-    page: state.NavigationReducer.page,
-  }
-}
-
-const mapDispatchToProps = (dispatch, ownProps) => {
-  return {
-    init: () => {
-    },
-    drawerToggle: (drawerOpen) => {
-      dispatch(drawerToggle(drawerOpen))
-    },
-    changeScreen: (page) => {
-      dispatch(changeScreen(page))
-      dispatch(drawerToggle(false))
-    }
-  }
-}
-
-const MenuContainer = connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(withStyles(styles)(Menu))
-export default MenuContainer
