+++ ui/src/components/ReadWriteCard.js
@@ -22,9 +22,6 @@ const styles = theme => ({
 });
 
 class ReadWriteCard extends React.Component {
-
-    
-
     render() {
     const { classes } = this.props;
         return (
